requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.1.0
tqdm>=4.66.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
